# Backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/team/{id}": {
      "get": {
        "operationId": "TeamController_getOneTeam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["team"]
      },
      "delete": {
        "operationId": "TeamController_deleteTeam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["team"]
      },
      "put": {
        "operationId": "TeamController_updateTeam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["team"]
      }
    },
...
  },
  "info": {
    "title": "Cricket Stats Backend",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}